Move useDropZone to compose
