+++ packages/components/src/drop-zone/style.scss
@@ -57,7 +57,3 @@
 .components-drop-zone__content-text {
 	font-family: $default-font;
 }
-
-.components-drop-zone__provider {
-	height: 100%;
-}
