+++ packages/components/src/index.js
@@ -48,10 +48,7 @@ export { default as __experimentalDimensionControl } from './dimension-control';
 export { default as Disabled } from './disabled';
 export { DisclosureContent as __unstableDisclosureContent } from './disclosure';
 export { default as Draggable } from './draggable';
-export {
-	default as DropZone,
-	useDropZone as __unstableUseDropZone,
-} from './drop-zone';
+export { default as DropZone } from './drop-zone';
 export { default as DropZoneProvider } from './drop-zone/provider';
 export { default as Dropdown } from './dropdown';
 export { default as DropdownMenu } from './dropdown-menu';
