+++ packages/compose/src/index.js
@@ -35,3 +35,4 @@ export { default as useDebounce } from './hooks/use-debounce';
 export { default as useThrottle } from './hooks/use-throttle';
 export { default as useMergeRefs } from './hooks/use-merge-refs';
 export { default as useRefEffect } from './hooks/use-ref-effect';
+export { default as __experimentalUseDropZone } from './hooks/use-drop-zone';
